requirements.txt
Flask==2.3.2
gunicorn==21.2.0
xmltodict==0.14.2
openpyxl==3.1.5
numpy==1.24.4
pandas==1.5.3
psycopg2-binary==2.9.9